RU.ini, US.ini, city_add.txt  added cities; geoCoord.pl bugfixes; relgui added Messjar checkbox (alpha); mutter2 bugfix in phase01.bin; fastjar stripped

+++ Astromaximum/mutter2/evclass.cpp
@@ -30,7 +30,7 @@ double Event::calcJD(long date)
 
 int Event::getDayNumber()
 {
-  return (julianDay-startJD)/MINUTE_STEP;
+  return (int)((julianDay-startJD)/MINUTE_STEP);
 }
 
 long Event::packDate(double date)
@@ -41,8 +41,8 @@ long Event::packDate(double date)
   now.tm_year=y-1900;
   now.tm_mon=m-1;
   now.tm_mday=d;
-  now.tm_hour=hour;
-  now.tm_min=(hour-now.tm_hour)*60;
+  now.tm_hour=(int)hour;
+  now.tm_min=(int)((hour-now.tm_hour)*60);
   now.tm_sec=0;
   now.tm_isdst = 0;
 #ifdef ANSITZ
